+++ test/Sentry.Tests/SentryLogTests.cs
@@ -22,6 +22,51 @@ public class SentryLogTests
     }
 
     [Fact]
+    public void Protocol_Default_VerifyAttributes()
+    {
+        var options = new SentryOptions
+        {
+            Environment = "my-environment",
+            Release = "my-release",
+        };
+        var sdk = new SdkVersion
+        {
+            Name = "Sentry.Test.SDK",
+            Version = "1.2.3-test+Sentry"
+        };
+
+        var log = new SentryLog(Timestamp, TraceId, (SentryLogLevel)24, "message")
+        {
+            Template = "template",
+            Parameters = ImmutableArray.Create<object>("params"),
+            ParentSpanId = ParentSpanId,
+        };
+        log.SetAttribute("attribute", "value");
+        log.SetDefaultAttributes(options, sdk);
+
+        log.Timestamp.Should().Be(Timestamp);
+        log.TraceId.Should().Be(TraceId);
+        log.Level.Should().Be((SentryLogLevel)24);
+        log.Message.Should().Be("message");
+        log.Template.Should().Be("template");
+        log.Parameters.Should().BeEquivalentTo(["params"]);
+        log.ParentSpanId.Should().Be(ParentSpanId);
+
+        log.TryGetAttribute("attribute", out object attribute).Should().BeTrue();
+        attribute.Should().Be("value");
+        log.TryGetAttribute("sentry.environment", out string environment).Should().BeTrue();
+        environment.Should().Be(options.Environment);
+        log.TryGetAttribute("sentry.release", out string release).Should().BeTrue();
+        release.Should().Be(options.Release);
+        log.TryGetAttribute("sentry.sdk.name", out string name).Should().BeTrue();
+        name.Should().Be(sdk.Name);
+        log.TryGetAttribute("sentry.sdk.version", out string version).Should().BeTrue();
+        version.Should().Be(sdk.Version);
+        log.TryGetAttribute("not-found", out object notFound).Should().BeFalse();
+        notFound.Should().BeNull();
+    }
+
+    [Fact]
     public void WriteTo_Envelope_MinimalSerializedSentryLog()
     {
         var options = new SentryOptions
@@ -31,7 +76,7 @@ public class SentryLogTests
         };
 
         var log = new SentryLog(Timestamp, TraceId, SentryLogLevel.Trace, "message");
-        log.SetDefaultAttributes(options);
+        log.SetDefaultAttributes(options, new SdkVersion());
 
         var envelope = Envelope.FromLog(log);
 
@@ -80,14 +125,6 @@ public class SentryLogTests
                 "sentry.release": {
                   "value": "my-release",
                   "type": "string"
-                },
-                "sentry.sdk.name": {
-                  "value": "{{SdkVersion.Instance.Name}}",
-                  "type": "string"
-                },
-                "sentry.sdk.version": {
-                  "value": "{{SdkVersion.Instance.Version}}",
-                  "type": "string"
                 }
               }
             }
@@ -117,7 +154,7 @@ public class SentryLogTests
         log.SetAttribute("boolean-attribute", true);
         log.SetAttribute("integer-attribute", 3);
         log.SetAttribute("double-attribute", 4.4);
-        log.SetDefaultAttributes(options);
+        log.SetDefaultAttributes(options, new SdkVersion { Name = "Sentry.Test.SDK", Version = "1.2.3-test+Sentry" });
 
         var envelope = EnvelopeItem.FromLog(log);
 
@@ -194,11 +231,11 @@ public class SentryLogTests
                   "type": "string"
                 },
                 "sentry.sdk.name": {
-                  "value": "{{SdkVersion.Instance.Name}}",
+                  "value": "Sentry.Test.SDK",
                   "type": "string"
                 },
                 "sentry.sdk.version": {
-                  "value": "{{SdkVersion.Instance.Version}}",
+                  "value": "1.2.3-test+Sentry",
                   "type": "string"
                 },
                 "sentry.trace.parent_span_id": {